+++ card.go
@@ -5,11 +5,8 @@
 package qiwi
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"log"
 	"net/url"
 	"strconv"
 	"time"
@@ -36,25 +33,8 @@ type CardsDetectResponse struct {
 
 // Detect detect card PS
 func (c *Cards) Detect(cardNumber string) (id int64, err error) {
-	body, err := c.client.makePostRequest(EndpointCardsDetect, url.Values{"cardNumber": {cardNumber}})
-	if err != nil {
-		return
-	}
-	defer body.Close()
-
-	bts, err := ioutil.ReadAll(body)
-	if err != nil {
-		return
-	}
-
-	buf := bytes.NewReader(bts)
-
-	log.Printf("%s", bts)
-
-	dec := json.NewDecoder(buf)
-
 	var r CardsDetectResponse
-	err = dec.Decode(&r)
+	err = c.client.makePostRequest(EndpointCardsDetect, &r, url.Values{"cardNumber": {cardNumber}})
 	if err != nil {
 		return
 	}
@@ -117,21 +97,10 @@ func (c *Cards) Payment(psID int64, amount float64, cardNumber string) (res Paym
 
 	endpoint := fmt.Sprintf(EndpointCardsPayment, psID)
 
-	body, err := c.client.makePostRequest(endpoint, req)
+	err = c.client.makePostRequest(endpoint, &res, req)
 	if err != nil {
 		return
 	}
-	defer body.Close()
-
-	bts, err := ioutil.ReadAll(body)
-	if err != nil {
-		return
-	}
-
-	log.Printf("%s\n", bts)
-
-	dec := json.NewDecoder(bytes.NewReader(bts))
-	err = dec.Decode(&res)
 
 	return
 }